+++ internal/objects/client_entity_writer.go
@@ -31,7 +31,7 @@ func (w *ClientEntityWriter) Create(object DRObject) {
 
 	w.Body.WriteByte(0x01) // Create
 
-	w.Body.WriteUInt16(object.RREntityProperties().ID) // Entity ID
+	w.Body.WriteUInt16(uint16(object.RREntityProperties().ID)) // Entity ID
 	w.Body.WriteByte(byte(GCObjectLookupTypeString))
 	w.Body.WriteCString(object.GetGCObject().GCType)
 
@@ -48,7 +48,7 @@ func (w *ClientEntityWriter) Init(object DRObject) {
 	w.dirty = true
 
 	w.Body.WriteByte(0x02) // Init
-	w.Body.WriteUInt16(object.RREntityProperties().ID)
+	w.Body.WriteUInt16(uint16(object.RREntityProperties().ID))
 
 	object.WriteInit(w.Body)
 }
@@ -56,9 +56,9 @@ func (w *ClientEntityWriter) Init(object DRObject) {
 func (w *ClientEntityWriter) CreateComponentAndInit(component DRObject, targetEntity DRObject) {
 	w.dirty = true
 
-	w.Body.WriteByte(0x32)                                   // Create Component
-	w.Body.WriteUInt16(targetEntity.RREntityProperties().ID) // Parent Entity ID
-	w.Body.WriteUInt16(component.RREntityProperties().ID)    // Component ID
+	w.Body.WriteByte(0x32)                                           // Create Component
+	w.Body.WriteUInt16(uint16(targetEntity.RREntityProperties().ID)) // Parent Entity ID
+	w.Body.WriteUInt16(uint16(component.RREntityProperties().ID))    // Component ID
 	w.Body.WriteByte(byte(GCObjectLookupTypeString))
 	w.Body.WriteCString(component.GetGCObject().GCType) // Component Type
 	w.Body.WriteByte(0x01)                              // Unk
@@ -69,8 +69,8 @@ func (w *ClientEntityWriter) CreateComponentAndInit(component DRObject, targetEn
 func (w *ClientEntityWriter) Update(object DRObject) {
 	w.dirty = true
 
-	w.Body.WriteByte(0x03)                             // MsgType Update
-	w.Body.WriteUInt16(object.RREntityProperties().ID) // Entity ID
+	w.Body.WriteByte(0x03)                                     // MsgType Update
+	w.Body.WriteUInt16(uint16(object.RREntityProperties().ID)) // Entity ID
 
 	object.WriteUpdate(w.Body)
 
@@ -82,8 +82,8 @@ func (w *ClientEntityWriter) Update(object DRObject) {
 func (w *ClientEntityWriter) BeginComponentUpdate(object DRObject) {
 	w.dirty = true
 
-	w.Body.WriteByte(0x35)                             // ComponentUpdate
-	w.Body.WriteUInt16(object.RREntityProperties().ID) // ComponentID
+	w.Body.WriteByte(0x35)                                     // ComponentUpdate
+	w.Body.WriteUInt16(uint16(object.RREntityProperties().ID)) // ComponentID
 }
 
 func (w *ClientEntityWriter) EndComponentUpdate(object DRObject) {
