+++ internal/objects/component_manipulators.go
@@ -27,7 +27,7 @@ func (n *Manipulators) WriteInit(b *byter.Byter) {
 	}
 }
 
-func (n *Manipulators) RemoveEquipmentByID(id uint16) {
+func (n *Manipulators) RemoveEquipmentByID(id uint32) {
 	toRemove := -1
 
 	for li, child := range n.Children() {
