+++ internal/objects/entity_player.go
@@ -106,7 +106,7 @@ func SendCreateNewPlayerEntity(rrplayer *RRPlayer, p *Player) {
 	clientEntityWriter.CreateComponentAndInit(manipulators, avatar)
 
 	equipment := avatar.GetChildByGCType("avatar.base.Equipment")
-	addCreateComponent(body, avatar.RREntityProperties().ID, equipment.RREntityProperties().ID, "avatar.base.Equipment")
+	addCreateComponent(body, uint16(avatar.RREntityProperties().ID), uint16(equipment.RREntityProperties().ID), "avatar.base.Equipment")
 
 	body.WriteByte(byte(len(equippedItems)))
 
@@ -180,7 +180,7 @@ func SendCreateNewPlayerEntity(rrplayer *RRPlayer, p *Player) {
 
 	// MODIFIERS //////////////////////////////////
 	// Modifiers are for modifying damage and defences
-	addCreateComponent(body, avatar.RREntityProperties().ID, NewID(), "Modifiers")
+	addCreateComponent(body, uint16(avatar.RREntityProperties().ID), NewID(), "Modifiers")
 
 	// Modifiers::readInit
 	body.WriteUInt32(0x00) //
@@ -200,7 +200,7 @@ func SendCreateNewPlayerEntity(rrplayer *RRPlayer, p *Player) {
 	//}
 
 	// SKILLS //////////////////////////////////
-	addCreateComponent(body, avatar.RREntityProperties().ID, NewID(), "avatar.base.skills")
+	addCreateComponent(body, uint16(avatar.RREntityProperties().ID), NewID(), "avatar.base.skills")
 
 	// Skills::readInit()
 	body.WriteUInt32(0xFFFFFFFF)
@@ -239,7 +239,7 @@ func SendCreateNewPlayerEntity(rrplayer *RRPlayer, p *Player) {
 	unitBehaviour := avatar.GetChildByGCNativeType("UnitBehavior")
 
 	if behaviorName == "avatar.base.UnitBehavior" {
-		addCreateComponent(body, avatar.RREntityProperties().ID, unitBehaviour.RREntityProperties().ID, "avatar.base.UnitBehavior")
+		addCreateComponent(body, uint16(avatar.RREntityProperties().ID), uint16(unitBehaviour.RREntityProperties().ID), "avatar.base.UnitBehavior")
 
 		behav := behavior.NewBehavior()
 		behav.Init(body, nil, nil)
@@ -290,7 +290,7 @@ func SendCreateNewPlayerEntity(rrplayer *RRPlayer, p *Player) {
 		body.WriteByte(0xFF)
 	} else {
 		// This is a monster behavior
-		addCreateComponent(body, avatar.RREntityProperties().ID, NewID(), "base.MeleeUnit.Behavior")
+		addCreateComponent(body, uint16(avatar.RREntityProperties().ID), NewID(), "base.MeleeUnit.Behavior")
 
 		behav := behavior.NewBehavior()
 		behav.Init(body, nil, nil)
@@ -345,7 +345,7 @@ func SendCreateNewPlayerEntity(rrplayer *RRPlayer, p *Player) {
 
 	// Init
 	body.WriteByte(0x02)
-	body.WriteUInt16(avatar.RREntityProperties().ID)
+	body.WriteUInt16(uint16(avatar.RREntityProperties().ID))
 
 	//WorldEntity::readInit
 	// Flags
