+++ internal/objects/entity_unit.go
@@ -26,7 +26,7 @@ func (n *Unit) WriteInit(b *byter.Byter) {
 
 	if n.UnitFlags&0x01 > 0 {
 		if n.RREntityProperties().OwnerID != 0 {
-			b.WriteUInt16(Players.Players[n.RREntityProperties().OwnerID].CurrentCharacter.RREntityProperties().ID) // Parent ID!!!!!
+			b.WriteUInt16(uint16(Players.Players[n.RREntityProperties().OwnerID].CurrentCharacter.RREntityProperties().ID)) // Parent ID!!!!!
 		} else {
 			b.WriteUInt16(0x00) // Parent ID!!!!!
 		}
